moved database util
